✨ improve(@roots/bud-imagemin): simplify api; swap out squoosh
