+++ sources/@roots/bud-imagemin/src/sharp/index.ts
@@ -1,0 +1,2 @@
+import {BudImageminSharp} from './sharp.js'
+export default BudImageminSharp
