+++ sources/@roots/bud-imagemin/src/svgo/index.ts
@@ -1,0 +1,2 @@
+import {BudImageminSvgo} from './svgo.js'
+export default BudImageminSvgo
