+++ sources/@roots/bud-imagemin/src/types.ts
@@ -1,6 +1,8 @@
 /// <reference types="@roots/bud" />
 
 import type {BudImageminExtension} from './extension.js'
+import type BudImageminSharp from './sharp/index.js'
+import type {BudImageminSvgo} from './svgo/svgo.js'
 
 declare module '@roots/bud-framework' {
   interface Bud {
@@ -9,5 +11,7 @@ declare module '@roots/bud-framework' {
 
   interface Modules {
     '@roots/bud-imagemin': BudImageminExtension
+    '@roots/bud-imagemin/sharp': BudImageminSharp
+    '@roots/bud-imagemin/svgo': BudImageminSvgo
   }
 }
